+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Write execution time for each test in the verbose mode.
 - When capture is disabled and verbose mode is on test names are printed
   twice: at the start and at the end with result.
+- `assert_error_msg_` assertions print return values if no error is generated
 
 ## 0.5.2